backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
neo4j==5.14.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# Search
elasticsearch==8.11.0

# NLP & ML
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
scispacy==0.5.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
bertopic==0.15.0
umap-learn==0.5.5
hdbscan==0.8.33

# OpenAI & LangChain
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
tiktoken==0.5.2

# Vector DB (choose one)
pinecone-client==2.2.4
# weaviate-client==3.25.3
# chromadb==0.4.18

# Data processing
pandas==2.1.3
numpy==1.26.2
biopython==1.81
requests==2.31.0
aiohttp==3.9.1

# API clients
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
